+++ src/renderer.rs
@@ -61,11 +61,14 @@ pub trait Renderer {
   /// ```rust
   /// # use gui::{BBox, Cap, Id, Renderer, Renderable};
   /// # use gui::derive::{Handleable, Widget};
+  /// # type Event = ();
   /// # #[derive(Debug, Widget, Handleable)]
+  /// # #[gui(Event = "Event")]
   /// # struct ConcreteWidget1 {
   /// #   id: Id,
   /// # }
   /// # #[derive(Debug, Widget, Handleable)]
+  /// # #[gui(Event = "Event")]
   /// # struct ConcreteWidget2 {
   /// #   id: Id,
   /// # }