+++ tests/common/mod.rs
@@ -33,11 +33,11 @@ use gui::Event;
 use gui::Handleable;
 use gui::Id;
 use gui::MutCap;
-use gui::UiEvent;
-use gui::UiEvents;
+use gui::UiEvents as GuiEvents;
 use gui::UnhandledEvent;
 use gui::UnhandledEvents;
 
+pub type UiEvents = GuiEvents<Event>;
 
 struct Handler<T>(T);
 
@@ -55,9 +55,9 @@ impl<T> Deref for Handler<T> {
   }
 }
 
-type EventFn = Fn(Id, Event, &mut MutCap) -> Option<UiEvents>;
-type CustomFn = Fn(Id, Box<Any>, &mut MutCap) -> Option<UiEvents>;
-type CustomRefFn = Fn(Id, &mut Any, &mut MutCap) -> Option<UiEvents>;
+type EventFn = Fn(Id, Event, &mut MutCap<Event>) -> Option<UiEvents>;
+type CustomFn = Fn(Id, Box<Any>, &mut MutCap<Event>) -> Option<UiEvents>;
+type CustomRefFn = Fn(Id, &mut Any, &mut MutCap<Event>) -> Option<UiEvents>;
 
 type EventHandler = Handler<Box<EventFn>>;
 type CustomHandler = Handler<Box<CustomFn>>;
@@ -85,7 +85,7 @@ impl TestWidgetBuilder {
   /// Set a handler for `Handleable::handle`.
   pub fn event_handler<F>(mut self, handler: F) -> Self
   where
-    F: 'static + Fn(Id, Event, &mut MutCap) -> Option<UiEvents>,
+    F: 'static + Fn(Id, Event, &mut MutCap<Event>) -> Option<UiEvents>,
   {
     self.event_handler = Some(Handler(Box::new(handler)));
     self
@@ -94,7 +94,7 @@ impl TestWidgetBuilder {
   /// Set a handler for `Handleable::handle_custom`.
   pub fn custom_handler<F>(mut self, handler: F) -> Self
   where
-    F: 'static + Fn(Id, Box<Any>, &mut MutCap) -> Option<UiEvents>,
+    F: 'static + Fn(Id, Box<Any>, &mut MutCap<Event>) -> Option<UiEvents>,
   {
     self.custom_handler = Some(Handler(Box::new(handler)));
     self
@@ -103,7 +103,7 @@ impl TestWidgetBuilder {
   /// Set a handler for `Handleable::handle_custom_ref`.
   pub fn custom_ref_handler<F>(mut self, handler: F) -> Self
   where
-    F: 'static + Fn(Id, &mut Any, &mut MutCap) -> Option<UiEvents>,
+    F: 'static + Fn(Id, &mut Any, &mut MutCap<Event>) -> Option<UiEvents>,
   {
     self.custom_ref_handler = Some(Handler(Box::new(handler)));
     self
@@ -122,6 +122,7 @@ impl TestWidgetBuilder {
 
 
 #[derive(Debug, Widget)]
+#[gui(Event = "Event")]
 pub struct TestWidget {
   id: Id,
   event_handler: Option<EventHandler>,
@@ -140,8 +141,8 @@ impl TestWidget {
   }
 }
 
-impl Handleable for TestWidget {
-  fn handle(&mut self, event: Event, cap: &mut MutCap) -> Option<UiEvents> {
+impl Handleable<Event> for TestWidget {
+  fn handle(&mut self, event: Event, cap: &mut MutCap<Event>) -> Option<UiEvents> {
     match self.event_handler.take() {
       Some(handler) => {
         let event = handler(self.id, event, cap);
@@ -152,18 +153,18 @@ impl Handleable for TestWidget {
     }
   }
 
-  fn handle_custom(&mut self, event: Box<Any>, cap: &mut MutCap) -> Option<UiEvents> {
+  fn handle_custom(&mut self, event: Box<Any>, cap: &mut MutCap<Event>) -> Option<UiEvents> {
     match self.custom_handler.take() {
       Some(handler) => {
         let event = handler(self.id, event, cap);
         self.custom_handler = Some(handler);
         event
       },
-      None => Some(UiEvent::Custom(event).into()),
+      None => Some(gui::UiEvent::Custom(event).into()),
     }
   }
 
-  fn handle_custom_ref(&mut self, event: &mut Any, cap: &mut MutCap) -> Option<UiEvents> {
+  fn handle_custom_ref(&mut self, event: &mut Any, cap: &mut MutCap<Event>) -> Option<UiEvents> {
     match self.custom_ref_handler.take() {
       Some(handler) => {
         let event = handler(self.id, event, cap);
@@ -176,8 +177,9 @@ impl Handleable for TestWidget {
 }
 
 #[allow(unused)]
-pub fn unwrap_custom<T>(events: UnhandledEvents) -> Box<T>
+pub fn unwrap_custom<E, T>(events: UnhandledEvents<E>) -> Box<T>
 where
+  E: Debug,
   T: 'static,
 {
   match events {